Send loaded features to zeus master on slave crash
If the slave crashes on first boot, zeus master needs to know what
files triggered that crash. Before, the master would just keep the
node in "crashed" state without tracking changes on any of its
features. Now, it waits for the slave to send the set of loaded files
over `local` and adds those to tracked features.

This fixes #321.