Fix "slave `boot` was killed", shorten backtraces
Extract command running from the `go` method. This ensures that the
process running the command never tries to re-register with the client
monitor on the Zeus side, which was what caused the "slave `boot` was
killed" messages mentioned in #410. This was originally broken by the
recursion elimination in #361 (turns out unrolling recursion is hard!).

As a pleasant side effect, this change also eliminates two more lines in
the `go` method from backtraces on commands.