Support rspec 3.0+ out of the box
Rspec 3.0 introduces separate _helper files for full tests and tests that
don't require the full rails app to be loaded. Because of this and because
the point of zeus is to pre-load your app, default to loading the full
rails_helper for zeus. You can override this behavior with the env var
RAILS_TEST_HELPER.

Fixes #474