Remove hacks around Minitest internals, make use of Minitest.autorun
Instead of trying to hack around Minitest internals, we'll simply hook into
the way that they internally call Minitest.run(ARGV) by replacing ARGV with
our desired test arguments. This is arguably a hack because we're mangling
ARGV but Minitest has executed with ARGV explicitly passed in for at least
two major releases so we'll go with it for now.