User bundler/setup instead of custom loading code
Adding "require 'bundler/setup'" to a file will add the bundled gems to
the load path. This is a much safer way to load gems instead of using
`gem <gemname>, <version>` where version is parsed from the Gemfile.

Loading 'bundler/setup.rb' does the following things:

1. Cleans load path, whose one function is to removing system gems from the
   load path.
2. Loads the gem specs from Gemfile.lock
3. Sets up the following path variables:
   BUNDLE_BIN_PATH
   BUNDLE_GEMFILE: location of the Gemfile
   PATH: Appends the location of `bundle` executable for the current
         Ruby version to the PATH.
   RUBYOPT
   RUBYLIB
3. Activate the loaded Gem file specs. This is where the command would
   throw an exception if there's a conflict in the case where a
   different version of a particular gem in the Gemfile.lock has already
   been loaded.

This makes the process more transparent. If there are any conflicts, an
exception would be raised and this can be caught and the resulting
backtrace can be handled properly.