Fix race in transition from UNBOOTED to BOOTING states
Zeus currently ignores restart messages when the slave process is in
the UNBOOTED state. If the slave process starts loading ruby files,
then one of those files changes and then Zeus completes the transition
to the BOOTING state, the changes won't trigger a restart. Instead,
Zeus should send a message to the slave process confirming that it
should start booting.