Track features from the backtrace of exceptions
When the action run by Zeus raises an exception, we report
`all_features` and crash. Since files do not get added to Ruby's
`$LOADED_FEATURES` list unless the are required without error, this
means that Zeus may not reload when you fix the underlying issue. This
is mostly plastered over by the fact that Zeus accumulates features
across all runs of a command so if your code loads once, it'll know
what to reload on if it crashes the next time. This doesn't work on
first boot-up and constrains some possible improvements to file
tracking in the SlaveNode. At Stripe, we're currently monkey patching
Kernel#require_relative in our Plan to work around this issue.

This changes `features_loaded_by` to catch exceptions in the executing
block and add the lines of the backtrace into the list of features
reported to Zeus. This should provide an accurate picture of the files
that were actually accessed leading to the crash. The existing shared
test helpers were stubbing out `Kernel#load`, breaking some
`LoadTracking` tests. I've moved helpers to be scoped to the modules
in which they're used and killed the shared file.