Report node as crashed if the process exits
Currently nodes do not monitor whether their processes exit.
A node will remain marked as "running" until it is reloaded, even
if the underlying process dies. Instead, it should be marked as crashed.
I believe this is true even if the underlying process exits with a
success code -- Zeus never expects its processes to die.