Refactor Zeus' core process management
This encapsulates all of the code for communicating directly
with Zeus' node processes into a `process` package. That's now
the only package that speaks over a unix socket to node processes.
The eventual goal is to completely replace the processtree package
with something more robust. This incremental step places a lot of
complexity inside a testable unit and should resolve at least a few
race conditions.

In order to accomplish this, I changed how node processes communicate back
to Zeus. Previously, when a new node started up it always communicated
using a single master socket. It now communicates using its parent's
command socket (which is actually how commands work already).