Handle errors reading from socket during boot gracefully
Currently when Zeus receives any error reading from the command socket
while waiting for a process to boot, it logs the error and then tries
to parse the message anyway. Since the message shouldn't be valid when
there's an error, we should instead enter an `SCrashed` state
immediately. It's we're throwing away useful debugging information
bufferred from an earlier read but I don't think that'll actually
happen in the existing Zeus code.